feat: Extract some methods into the class to optimize the code structure
