+++ Plot.Core/EventProcess/IPlotEvent.cs
@@ -1,7 +1,9 @@
+using Plot.Core.Renderables.Axes;
+
 namespace Plot.Core.EventProcess
 {
     public interface IPlotEvent
     {
-        void Process();
+        void Process(AxisManager axisManager);
     }
 }
