+++ Plot.Core/EventProcess/MouseScrollEvent.cs
@@ -1,9 +1,11 @@
+using Plot.Core.Renderables.Axes;
+
 namespace Plot.Core.EventProcess
 {
     public class MouseScrollEvent : IPlotEvent
     {
         private readonly EventManager m_manager;
-        private InputState m_inputState;
+        private readonly InputState m_inputState;
 
         private readonly float m_frac = 0.15f;
 
@@ -13,7 +15,7 @@ namespace Plot.Core.EventProcess
             m_inputState = inputState;
         }
 
-        public void Process()
+        public void Process(AxisManager axisManager)
         {
             float increment = 1.0f + m_frac;
             float decrement = 1.0f - m_frac;
@@ -22,7 +24,7 @@ namespace Plot.Core.EventProcess
             float xFrac = m_inputState.m_scrollUp ? increment : decrement;
             float yFrac = m_inputState.m_scrollUp ? increment : decrement;
 
-            m_manager.ZoomCenter(xFrac, yFrac, m_inputState.m_x, m_inputState.m_y);
+            axisManager.ZoomByFrac(xFrac, yFrac, m_inputState.m_x, m_inputState.m_y);
         }
     }
 }
