+++ Plot.Core/Series/IPlotSeries.cs
@@ -7,20 +7,15 @@ namespace Plot.Core.Series
     {
         Axis XAxis { get; }
         Axis YAxis { get; }
-        Figure Figure { get; }
 
         Color Color { get; set; }
         float LineWidth { get; set; }
         string Label { get; set; }
-
-        float Scale { get; set; }
 
         AxisLimits GetAxisLimits();
 
         void ValidateData();
 
         void Plot(Bitmap bmp, bool lowQuality, float scale);
-
-        PlotDimensions Dims { get; }
     }
 }
